% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calculate_duration}
\alias{calculate_duration}
\title{Calculate outbreak duration from a single scenario simulation}
\usage{
calculate_duration(df)
}
\arguments{
\item{df}{A data.frame of a single simulation}
}
\value{
The integer number of generations until the outbreak went extinct.
}
\description{
Calculate outbreak duration from a single scenario simulation
}
\author{
Sophie Meakin
}
