% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calculate_pep}
\alias{calculate_pep}
\title{Count number of individuals who received PEP}
\usage{
calculate_pep(df)
}
\arguments{
\item{df}{A data.frame of a single simulation}
}
\value{
The integer number of individuals who received PEP.
}
\description{
Count number of individuals who received PEP
}
\author{
Sophie Meakin
}
