% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{calculate_r0}
\alias{calculate_r0}
\title{Calculate R0 from single scenario simulation}
\usage{
calculate_r0(df)
}
\arguments{
\item{df}{A data.frame of a single simulation}
}
\value{
Numeric distribution of number of secondary cases in simulation.
}
\description{
Calculate R0 from single scenario simulation
}
\author{
Sophie Meakin
}
