% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run PEP branching process model}
\usage{
outbreak_model(
  prop_pep,
  perc_risk,
  hrc_mu,
  hrc_disp,
  lrc_mu,
  lrc_disp,
  p_hrc_case,
  p_lrc_case,
  n_initialcases,
  cap_max_gen,
  cap_max_cases
)
}
\arguments{
\item{prop_pep}{Proportion of high-risk contacts that receive PEP}

\item{perc_risk}{Percentage reduction in risk of becoming a case after receiving PEP}

\item{hrc_mu}{Average number of high-risk contacts}

\item{hrc_disp}{Dispersion parameter for number of high-risk contacts}

\item{lrc_mu}{Average number of low-risk contacts}

\item{lrc_disp}{Dispersion parameter for number of low-risk contacts}

\item{p_hrc_case}{Probability that a high-risk contact becomes a case}

\item{p_lrc_case}{Probability that a low-risk contact becomes a case}

\item{n_initialcases}{Integer number of initial cases}

\item{cap_max_gen}{Stop the simulation when this many generations is reached}

\item{cap_max_cases}{Stop the simulation when this many cases is reached}
}
\value{
A data.frame of cases by generation, with columns:
\itemize{
\item generation
\item case_id
\item case_type: one of "PEP", "HRC" (no PEP), "LRC".
\item infector: case_id of infector (currently not used)
\item new_cases: total new cases
\item new_cases_hrc: new cases that were high-risk contacts
\item new_cases_pep: new cases that were high-risk contacts that received PEP
\item new_cases_lrc: new cases that were low-risk contacts
\item n_hrc: number of high-risk contacts that did not receive PEP
\item n_pep: number of high-risk contacts that received PEP
\item n_lrc: number of low-risk contacts
}
}
\description{
Run a single instance of the branching process model for post-exposure prophylaxis for high-risk contacts. See Details for a full description of branching process model.
}
\details{
Each case has some number of high- and low-risk contacts, each defined by a negative binomial distribution.
A fixed proportion \code{prop_pep} of high-risk contacts receive PEP.
High-risk contacts who do not receive PEP become cases with probability \code{p_hrc_case}; high-risk contacts who receive PEP become cases with lower probability \code{p_hrc_case*perc_risk}.
Low-risk contacts do not receive PEP, and become cases with probability \code{p_lrc_case}.
}
\author{
Sophie Meakin, Joel Hellewell
}
