% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_sweep.R
\name{parameter_sweep}
\alias{parameter_sweep}
\title{Sweep across parameters}
\usage{
parameter_sweep(
  scenarios = NULL,
  n_sim = 1,
  sim_fn = NULL,
  show_progress = TRUE
)
}
\arguments{
\item{scenarios}{A data.frame containing all gridded scenarios.}

\item{n_sim}{The number of simulations to make per scenario. Numeric, defaults to 1.}

\item{sim_fn}{The vectorised model simulation function.}

\item{show_progress}{Show the progress of the parameter sweep. Logical, defaults to \code{TRUE}.}
}
\value{
A nested tibble containing the parameters for each scenario and a nested list of output
from \code{sim_fn}.
}
\description{
Explore scenarios using gridding with sampling for parameters not in the grid. Parameters that
are included in the grid are currently hard coded. Use the \code{future} package to control parallisation
outside of the function.
}
\author{
Sam Abbott, Sophie Meakin, Joel Hellewell
}
