% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_r0}
\alias{plot_r0}
\title{Visualise change in R0}
\usage{
plot_r0(sweep_summary, p)
}
\arguments{
\item{sweep_summary}{Output from summarise_results().}

\item{p}{Proportion of HRCs that receive PEP.}
}
\value{
A ggplot object.
}
\description{
Visualise relative of risk of becoming a case versus average number of secondary cases for a \% of HRCs that receive PEP (\code{p}).
}
\author{
Sophie Meakin
}
