% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_sim.R
\name{scenario_sim}
\alias{scenario_sim}
\title{Run a specified number of simulations with identical parameters.}
\usage{
scenario_sim(
  n_sim,
  prop_pep,
  rel_risk,
  hrc_mu,
  hrc_disp,
  lrc_mu,
  lrc_disp,
  p_hrc_case,
  p_lrc_case,
  n_initialcases,
  cap_max_gen,
  cap_max_cases
)
}
\arguments{
\item{prop_pep}{Proportion of high-risk contacts that receive PEP}

\item{rel_risk}{Relative risk of a high-risk contact becoming a case after receiving PEP}

\item{hrc_mu}{Average number of high-risk contacts}

\item{hrc_disp}{Dispersion parameter for number of high-risk contacts}

\item{lrc_mu}{Average number of low-risk contacts}

\item{lrc_disp}{Dispersion parameter for number of low-risk contacts}

\item{p_hrc_case}{Probability that a high-risk contact becomes a case}

\item{p_lrc_case}{Probability that a low-risk contact becomes a case}

\item{n_initialcases}{Integer number of initial cases}

\item{cap_max_gen}{Stop the simulation when this many generations is reached}

\item{cap_max_cases}{Stop the simulation when this many cases is reached}

\item{nsim}{Number of simulations to run}
}
\value{
A data.frame containing the results for multiple simulations.
}
\description{
Run a specified number of simulations with identical parameters.
}
\author{
Joel Hellewell, Sophie Meakin
}
