% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_model.R
\name{outbreak_model}
\alias{outbreak_model}
\title{Run a single instance of the branching process model.}
\usage{
outbreak_model(
  prop_pep,
  perc_risk,
  hrc_mu,
  hrc_disp,
  lrc_mu,
  lrc_disp,
  p_hrc_case,
  p_lrc_case,
  n_initialcases,
  cap_max_gen,
  cap_max_cases
)
}
\arguments{
\item{prop_pep}{Proportion of high-risk contacts that receive PEP}

\item{perc_risk}{Percentage reduction in risk of becoming a case after receiving PEP}

\item{hrc_mu}{Average number of high-risk contacts}

\item{hrc_disp}{Dispersion parameter for number of high-risk contacts}

\item{lrc_mu}{Average number of low-risk contacts}

\item{lrc_disp}{Dispersion parameter for number of low-risk contacts}

\item{p_hrc_case}{Probability that a high-risk contact becomes a case}

\item{p_lrc_case}{Probability that a low-risk contact becomes a case}

\item{n_initialcases}{Integer number of initial cases}

\item{cap_max_gen}{Stop the simulation when this many generations is reached}

\item{cap_max_cases}{Stop the simulation when this many cases is reached}
}
\value{
A data.frame of high- and low-risk contacts and cases by generation.
}
\description{
Run a single instance of the branching process model.
}
\author{
Joel Hellewell, Sophie Meakin
}
