% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_setup.R
\name{outbreak_setup}
\alias{outbreak_setup}
\title{Set up initial cases for branching process}
\usage{
outbreak_setup(n_initialcases)
}
\arguments{
\item{n_initialcases}{Integer number of initial cases}
}
\value{
A data.frame of initial cases in outbreak with columns. Passed to pepbp::outbreak_step().
}
\description{
Set up initial cases for branching process
}
\author{
Joel Hellewell, Sophie Meakin
}
