% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_step.R
\name{outbreak_step}
\alias{outbreak_step}
\title{Move branching process forward by one generation}
\usage{
outbreak_step(
  outbreak_data,
  hrc_mu,
  hrc_disp,
  lrc_mu,
  lrc_disp,
  p_hrc_case,
  p_lrc_case
)
}
\arguments{
\item{outbreak_data}{A data.frame of cases}
}
\value{
A data.frame updated with one new generation of cases.
}
\description{
Move branching process forward by one generation
}
\author{
Joel Hellewell, Sophie Meakin
}
